fix: Creator info gets deleted after download banner page
