+++ apps/main-landing/src/app/creators/components/banner/assets/Banner_2.png
