+++ apps/main-landing/src/app/creators/components/banner/assets/Banner_7.png
