+++ apps/main-landing/src/app/creators/components/banner/assets/index.ts
