+++ apps/main-landing/src/app/creators/components/banner/index.ts
@@ -1,0 +1,1 @@
+export { Banner } from './banner';
